requirements.txt
a2wsgi==1.10.10
agate==1.9.1
aiologic==0.14.0
aiosmtplib==4.0.2
aiosqlite==0.21.0
alembic==1.16.5
annotated-types==0.7.0
anyio==4.11.0
apache-airflow==3.0.0
apache-airflow-core==3.0.0
apache-airflow-providers-common-compat==1.7.4
apache-airflow-providers-common-io==1.6.3
apache-airflow-providers-common-sql==1.28.1
apache-airflow-providers-smtp==2.3.0
apache-airflow-providers-standard==1.8.0
apache-airflow-task-sdk==1.0.0
argcomplete==3.6.2
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.9.2
asn1crypto==1.5.1
attrs==25.3.0
authlib==1.6.5
babel==2.17.0
backports-tarfile==1.2.0
black==25.9.0
blinker==1.9.0
boto3==1.40.59
botocore==1.40.59
cachetools==6.2.1
cadwyn==5.4.4
certifi==2025.8.3
cffi==1.17.1
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
colorlog==6.9.0
confluent-kafka==2.12.2
coverage==7.11.0
cron-descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.0
daff==1.4.2
dbt-adapters==1.17.3
dbt-artifacts-parser==0.11.0
dbt-common==1.33.0
dbt-core==1.10.13
dbt-extractor==0.6.0
dbt-protos==1.0.382
dbt-semantic-interfaces==0.9.0
dbt-snowflake==1.9.2
dbterd==1.21.0
deepdiff==8.6.1
deprecated==1.2.18
diff-cover==9.7.1
dill==0.4.0
distlib==0.4.0
dnspython==2.8.0
dotenv==0.9.9
email-validator==2.3.0
faker==37.6.0
fastapi==0.117.1
fastapi-cli==0.0.13
fastapi-cloud-cli==0.2.0
fastavro==1.12.1
filelock==3.20.0
flask==3.1.2
fsspec==2025.9.0
gitdb==4.0.12
gitpython==3.1.45
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.15
idna==3.10
importlib-metadata==8.7.0
iniconfig==2.1.0
isodate==0.6.1
itsdangerous==2.2.0
jaraco-classes==3.4.0
jaraco-context==6.0.1
jaraco-functools==4.3.0
jinja2==3.1.6
jinja2-simple-tags==0.6.1
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kagglehub==0.3.13
keyring==25.6.0
lazy-object-proxy==1.12.0
leather==0.4.0
libcst==1.8.4
linkify-it-py==2.0.3
lockfile==0.12.2
mako==1.3.10
markdown-it-py==4.0.0
markupsafe==3.0.2
mashumaro==3.14
mdurl==0.1.2
methodtools==0.4.7
minio==7.2.18
more-itertools==10.8.0
msgpack==1.1.2
msgspec==0.19.0
mypy-extensions==1.1.0
mysql-connector-python==9.5.0
networkx==3.5
nodeenv==1.9.1
numpy==2.3.3
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orderly-set==5.5.0
packaging==25.0
pandas==2.3.2
parsedatetime==2.6
pathspec==0.12.1
pendulum==3.1.0
platformdirs==4.5.0
pluggy==1.6.0
polars==1.33.1
pre-commit==4.3.0
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.11
pyarrow==22.0.0
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic-core==2.33.2
pyflink==1.0
pygments==2.19.2
pyjwt==2.10.1
pymysql==1.1.2
pyopenssl==25.3.0
pytest==8.4.2
pytest-cov==7.0.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-slugify==8.0.4
pytimeparse==1.1.8
pytokens==0.2.0
pytz==2025.2
pywin32-ctypes==0.2.3
pyyaml==6.0.2
referencing==0.36.2
regex==2025.10.23
requests==2.32.5
retryhttp==1.3.3
rich==14.1.0
rich-argparse==1.7.1
rich-toolkit==0.15.1
rignore==0.6.4
rpds-py==0.27.1
ruff==0.14.2
s3transfer==0.14.0
sentry-sdk==2.38.0
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake==1.8.0
snowflake-connector-python==3.18.0
snowflake-core==1.8.0
snowflake-legacy==1.0.2
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
sqlalchemy==1.4.54
sqlalchemy-jsonfield==1.0.2
sqlalchemy-utils==0.42.0
sqlfluff==3.5.0
sqlfluff-templater-dbt==3.5.0
sqlparse==0.5.3
starlette==0.48.0
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tblib==3.2.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
tomlkit==0.13.3
tqdm==4.67.1
typer==0.19.2
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
uc-micro-py==1.0.3
universal-pathlib==0.2.6
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.37.0
virtualenv==20.35.3
watchfiles==1.1.0
websockets==15.0.1
werkzeug==3.1.3
wirerope==1.0.0
wrapt==1.17.3
zipp==3.23.0
